requirements.txt
# Core Dependencies
PyMuPDF==1.23.14
pypdf==3.17.4
python-dotenv==1.0.0

# AI API Clients
openai==1.12.0
mistralai==0.1.8

# HTTP and Networking
httpx==0.26.0
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
pathlib2==2.3.7
typing-extensions==4.9.0

# Development and Testing (Optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

# Logging and Monitoring
structlog==23.2.0 